projects: adalm-lsmspg: Add support for max32655fthr Both tinyiiod and curvetrace_examples.
