+++ projects/adalm-lsmspg/boards/curvetrace_example/max32655fthr.conf
@@ -1,0 +1,1 @@
+# max32655fthr uses standard UART (no USB UART)
